requirements.txt
boto3
xmltodict
djangorestframework
markdown
requests
django-cors-headers

-r submodules/corpus-viewer/requirements.txt